Avoid a warning when printing a pid_t. Pointed out by Morten Welinder.
        * glib/gslice.c (mem_error): Avoid a warning when printing a pid_t.
        Pointed out by Morten Welinder.


svn path=/trunk/; revision=7196

+++ glib/gslice.c
@@ -1189,7 +1189,7 @@ mem_error (const char *format,
   /* at least, put out "MEMORY-ERROR", in case we segfault during the rest of the function */
   fputs ("\n***MEMORY-ERROR***: ", stderr);
   pname = g_get_prgname();
-  fprintf (stderr, "%s[%u]: GSlice: ", pname ? pname : "", getpid());
+  fprintf (stderr, "%s[%ld]: GSlice: ", pname ? pname : "", (long)getpid());
   va_start (args, format);
   vfprintf (stderr, format, args);
   va_end (args);
